Restructure to match rest of app.
