+++ client/src/components/test-environment/TestEnvironment.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Table } from "react-bootstrap";
-import Line from "../shared/line";
+import Line from "../shared/Line";
 import { useSelector } from "react-redux";
 
 const TestEnvironment = () => {
